Split predict Fstat into its own module
